Massive rewrite of geometry rendering system.  All is prepared for cluster-groups.  All it needs now really is to have a list of Clusters that are read in as DrawableObjects and isCluster set.  Then a loop at draw-time that renders the draw list in multiple places according to the attached plClusterGroup key.
Note:  I needed to do some hacks to plClusterGroup and plSpanInstance so I had access to some protected.  This code will not compile unless you move what it needs to public (it'll show obvious errors).  If you have a better way to get the Cluster data read in feel free to fix it.

git-svn-id: http://uru.zrax.net/svn/PlasmaViewer@22 46d49313-71c2-4fea-ab98-6c881a270271